+++ testdocker/cli/__init__.py
@@ -1,0 +1,12 @@
+"""
+testdocker.cli
+~~~~~~~~~~~~~~
+
+CLI interface package for testdocker.
+
+:copyright: (c) 2017 by Joe Black.
+:license: Apache2.
+
+"""
+
+from . import main